requirements.txt
blessings==1.7
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
dlib==19.17.0
dominate==2.3.5
gpustat==0.6.0
idna==2.8
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.16.4
nvidia-ml-py3==7.352.0
opencv-python==4.1.0.25
Pillow==6.1.0
pkg-resources==0.0.0
psutil==5.6.3
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.0.2
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchfile==0.1.0
torchvision==0.3.0
tornado==6.0.3
urllib3==1.25.3
visdom==0.1.8.8
websocket-client==0.56.0
